requirements.txt
# Core scientific stack
numpy>=1.24
scipy>=1.11
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.3

# Deep learning (CUDA/MPS/CPU autodetected by PyTorch at runtime)
torch>=2.2

# EEG handling
mne>=1.5           # friendly EDF reader + signal utils
pyedflib>=0.1.35   # EDF backend used by MNE

# Visualization (EDA)
matplotlib>=3.8
tensorboard>=2.16
